move re-exports into new module, and use libc instead of std::os::raw
next step is to re-write the API module with safe(r), idiomatic rust.
Motto: "make invalid states unrepresentable"
