+++ kdbplus/Cargo.toml
@@ -26,6 +26,7 @@ tokio = { version = "1", features = [
 tokio-native-tls = { version = "0.3", optional = true }
 async-trait = { version = "0.1", optional = true }
 async-recursion = "1"
+libc = "0.2"
 sha1_smol = { version = "1", optional = true }
 
 [[test]]
